caption_generation/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
Pillow>=10.0.0
rasterio>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
peft>=0.6.0
safetensors>=0.4.0
sentencepiece>=0.1.99
pyyaml>=6.0
jsonlines>=4.0.0
openai-clip